refactor(error): パスキーのエラー写像を認証エラーから分離する
パスキーのハンドラが認証(パスワード/セッション)エラーまで含む広い対応表を参照していた
のを、パスキー専用の対応表へ分離する。

- passkey-error.ts に PasskeyRegistrationError / PasskeyVerificationError の写像を切り出す
- auth-error.ts はパスワード・セッション由来のエラー(InvalidCredentials / UserAlreadyExists /
  NoSession)のみを扱う
- PasskeyClient は Passkey* / Unexpected しか返さないため、分離しても挙動は不変
