+++ application/apps/web/src/server/error/passkey-error.ts
@@ -1,14 +1,9 @@
-import {
-  InvalidCredentialsError,
-  NoSessionError,
-  UserAlreadyExistsError,
-} from '@trend-diary/authentication'
+import { PasskeyRegistrationError, PasskeyVerificationError } from '@trend-diary/authentication'
 import throwHttpErrorByTable, { type ErrorStatusTable } from './throw-http-error'
 
 const ERROR_STATUS_TABLE: ErrorStatusTable = [
-  [InvalidCredentialsError, 401],
-  [UserAlreadyExistsError, 409],
-  [NoSessionError, 401],
+  [PasskeyRegistrationError, 400],
+  [PasskeyVerificationError, 401],
 ]
 
 export default function throwHttpError(error: Error): never {
